Add bulk method
This is a faster but less precise method of generating the data,
which is useful when generating huge amounts of data.

Instead of generating literals directly using lit_exp, for each
BLOCK_SIZE bytes, generate a SAMPLE_SIZE buffer of literals following
the desired distribution, and then sample randomly from this buffer.

Increasing the sample size improves the precision but reduces speed.
As the sample is renewed for each block, the overall distribution
improves with the number of blocks generated. The current values are
a tradeof.